+++ src/components/tasks/Tasks.css
@@ -1,5 +1,5 @@
 .task {
-    background: #fff;
+    background:lightgray;
     text-align: start;
     line-height: 2;
     margin: 5rem;
